+++ src/profile-logic/marker-tree.js
@@ -1,0 +1,114 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
+
+// @flow
+
+import { getMarkerFullDescription, getMarkerCategory } from './marker-data';
+
+import { formatSeconds } from '../utils/format-numbers';
+
+import type { Milliseconds } from '../types/units';
+import type { Marker, MarkerIndex } from '../types/profile-derived';
+
+export type MarkerDisplayData = {|
+  start: string,
+  duration: string,
+  name: string,
+  category: string,
+|};
+
+class MarkerTree {
+  _getMarker: MarkerIndex => Marker;
+  _markerIndexes: MarkerIndex[];
+  _zeroAt: Milliseconds;
+  _displayDataByIndex: Map<MarkerIndex, MarkerDisplayData>;
+
+  constructor(
+    getMarker: MarkerIndex => Marker,
+    markerIndexes: MarkerIndex[],
+    zeroAt: Milliseconds
+  ) {
+    this._getMarker = getMarker;
+    this._markerIndexes = markerIndexes;
+    this._zeroAt = zeroAt;
+    this._displayDataByIndex = new Map();
+  }
+
+  getRoots(): MarkerIndex[] {
+    return this._markerIndexes;
+  }
+
+  getChildren(markerIndex: MarkerIndex): MarkerIndex[] {
+    return markerIndex === -1 ? this.getRoots() : [];
+  }
+
+  hasChildren(_markerIndex: MarkerIndex): boolean {
+    return false;
+  }
+
+  getAllDescendants() {
+    return new Set();
+  }
+
+  getParent(): MarkerIndex {
+    // -1 isn't used, but needs to be compatible with the call tree.
+    return -1;
+  }
+
+  getDepth() {
+    return 0;
+  }
+
+  hasSameNodeIds(tree) {
+    return this._markerIndexes === tree._markerIndexes;
+  }
+
+  getDisplayData(markerIndex: MarkerIndex): MarkerDisplayData {
+    let displayData = this._displayDataByIndex.get(markerIndex);
+    if (displayData === undefined) {
+      const marker = this._getMarker(markerIndex);
+      const name = getMarkerFullDescription(marker);
+      const category = getMarkerCategory(marker);
+
+      displayData = {
+        start: _formatStart(marker.start, this._zeroAt),
+        duration: marker.incomplete ? 'unknown' : _formatDuration(marker.dur),
+        name,
+        category,
+      };
+      this._displayDataByIndex.set(markerIndex, displayData);
+    }
+    return displayData;
+  }
+}
+
+function _formatStart(start: number, zeroAt) {
+  return formatSeconds(start - zeroAt);
+}
+
+function _formatDuration(duration: number): string {
+  if (duration === 0) {
+    return '—';
+  }
+  let maximumFractionDigits = 1;
+  if (duration < 0.01) {
+    maximumFractionDigits = 3;
+  } else if (duration < 1) {
+    maximumFractionDigits = 2;
+  }
+  return (
+    duration.toLocaleString(undefined, {
+      minimumFractionDigits: 0,
+      maximumFractionDigits,
+    }) + 'ms'
+  );
+}
+
+export function getMarkerTree(
+  getMarker: MarkerIndex => Marker,
+  markerIndexes: MarkerIndex[],
+  zeroAt: Milliseconds
+): MarkerTree {
+  return new MarkerTree(getMarker, markerIndexes, zeroAt);
+}
